Refactor: - Merged powerConnections.cs into Plugin.cs as it's only used there. - Added "Mechtrauma" identifier to all hooks to avoid conflicts. - Moved all components to the Mechtrauma namespace. - Made all Mechtrauma components public.
