+++ Mechtrauma/Lua/Scripts/Shared/mechtraumaFunctions.lua
@@ -2,7 +2,7 @@ MT.F = {}
 CentralComputer = {}
 CentralComputer.online = true
 
-LuaUserData.RegisterTypeBarotrauma("Items.Components.SimpleGenerator")
+-- LuaUserData.RegisterTypeBarotrauma("Items.Components.SimpleGenerator")
 
 -- Hull:Condition ratio for oxygen is 2333:1 and a player breaths 700 oxygen per second. 
 -- human breaths 700 oxygen/second and that requires to 0.3
