+++ SharedProject/SharedSource/ControllablePowerContainer.cs
@@ -12,9 +12,9 @@ using Microsoft.Xna.Framework;
 using Barotrauma.Items.Components;
 using System.Linq;
 
-namespace Barotrauma.Items.Components 
+namespace Mechtrauma 
 {
-    class ControllablePowerContainer : PowerContainer {
+    public class ControllablePowerContainer : PowerContainer {
 
         [Editable, Serialize(true, IsPropertySaveable.Yes, description: "Turn the power container output on and off", alwaysUseInstanceValues: true)]
         public bool IsOn {
